shutter: move proto pkg to internal
