+++ txnprovider/shutter/decryption_keys_validator_test.go
@@ -27,8 +27,8 @@ import (
 	"github.com/erigontech/erigon-lib/log/v3"
 	"github.com/erigontech/erigon/turbo/testlog"
 	"github.com/erigontech/erigon/txnprovider/shutter"
+	shutterproto "github.com/erigontech/erigon/txnprovider/shutter/internal/proto"
 	"github.com/erigontech/erigon/txnprovider/shutter/internal/testhelpers"
-	shutterproto "github.com/erigontech/erigon/txnprovider/shutter/proto"
 )
 
 func TestDecryptionKeysValidator(t *testing.T) {
