+++ txnprovider/shutter/internal/proto/serialization.go
